# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShakaSite Construction Timesheet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SeedResult {
  message: string;
  seeded: boolean;
}

export interface Worker {
  id: number;
  name: string;
  /** worker or manager */
  role: string;
  /** @nullable */
  mobileNumber?: string | null;
  daysPerWeek: number;
  /** @nullable */
  siteManagerId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWorkerBody {
  name: string;
  role: string;
  /** @nullable */
  mobileNumber?: string | null;
  daysPerWeek: number;
  /** @nullable */
  siteManagerId?: number | null;
}

export interface UpdateWorkerBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  daysPerWeek?: number | null;
  /** @nullable */
  siteManagerId?: number | null;
}

export interface Worksite {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateWorksiteBody {
  name: string;
  /** @nullable */
  address?: string | null;
  isActive: boolean;
}

export interface Job {
  id: number;
  name: string;
  worksiteId: number;
  /** @nullable */
  budgetedHours?: number | null;
  hoursUsed: number;
  /** @nullable */
  deadline?: string | null;
  /** active, completed, on_hold */
  status: string;
  createdAt: string;
}

export interface Milestone {
  id: number;
  jobId: number;
  name: string;
  /** @nullable */
  hoursAtMilestone?: number | null;
  /** @nullable */
  completedAt?: string | null;
  sortOrder: number;
}

export interface JobDetail {
  id: number;
  name: string;
  worksiteId: number;
  /** @nullable */
  budgetedHours?: number | null;
  hoursUsed: number;
  /** @nullable */
  deadline?: string | null;
  status: string;
  milestones: Milestone[];
  createdAt: string;
}

export interface CreateJobBody {
  name: string;
  worksiteId: number;
  /** @nullable */
  budgetedHours?: number | null;
  /** @nullable */
  deadline?: string | null;
  status?: string;
}

export interface UpdateJobBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  budgetedHours?: number | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface CreateMilestoneBody {
  name: string;
  /** @nullable */
  hoursAtMilestone?: number | null;
  sortOrder: number;
}

export interface TimesheetEntry {
  id: number;
  workerId: number;
  worksiteId: number;
  /** @nullable */
  jobId?: number | null;
  /** ISO date YYYY-MM-DD */
  date: string;
  /**
   * HH:MM format
   * @nullable
   */
  startTime?: string | null;
  /**
   * HH:MM format
   * @nullable
   */
  finishTime?: string | null;
  /** @nullable */
  lunchMinutes?: number | null;
  /** @nullable */
  totalHours?: number | null;
  /**
   * JSON string of task type hours
   * @nullable
   */
  taskBreakdown?: string | null;
  /** @nullable */
  notes?: string | null;
  isSignedOff: boolean;
  /** @nullable */
  signedOffBy?: number | null;
  /** @nullable */
  signedOffAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTimesheetEntryBody {
  workerId: number;
  worksiteId: number;
  /** @nullable */
  jobId?: number | null;
  date: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  finishTime?: string | null;
  /** @nullable */
  lunchMinutes?: number | null;
  /** @nullable */
  totalHours?: number | null;
  /** @nullable */
  taskBreakdown?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateTimesheetEntryBody {
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  finishTime?: string | null;
  /** @nullable */
  lunchMinutes?: number | null;
  /** @nullable */
  totalHours?: number | null;
  /** @nullable */
  taskBreakdown?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  jobId?: number | null;
}

export interface SignOffBody {
  signedOffBy: number;
}

export interface SignOffWeekBody {
  workerId: number;
  weekStart: string;
  signedOffBy: number;
}

export interface Message {
  id: number;
  fromWorkerId: number;
  toWorkerId: number;
  body: string;
  isRead: boolean;
  createdAt: string;
}

export interface SendMessageBody {
  fromWorkerId: number;
  toWorkerId: number;
  body: string;
}

export interface ExportTimesheetBody {
  /** @nullable */
  workerId?: number | null;
  /** @nullable */
  weekStart?: string | null;
  /** pdf or excel */
  format: string;
}

export interface ExportResult {
  url: string;
  filename: string;
  format: string;
}

export interface WorkerStats {
  workerId: number;
  workerName: string;
  weeklyHours: number;
  submittedDaysThisWeek: number;
  daysPerWeek: number;
  streakDays: number;
  /** @nullable */
  topWorksite?: string | null;
  /** @nullable */
  topTaskType?: string | null;
  missedYesterday: boolean;
  recentEntries: TimesheetEntry[];
}

export interface JobSummary {
  id: number;
  name: string;
  worksiteName: string;
  /** @nullable */
  budgetedHours?: number | null;
  hoursUsed: number;
  /** @nullable */
  budgetPercent?: number | null;
  /** @nullable */
  deadline?: string | null;
  status: string;
  isAtRisk: boolean;
  isOverBudget: boolean;
  workerCount: number;
}

export interface ManagerStats {
  activeJobsCount: number;
  totalWorkersThisWeek: number;
  totalHoursThisWeek: number;
  pendingSignoffs: number;
  jobsOverBudget: number;
  /** Jobs at 70% or more of budget */
  jobsAtRisk: number;
  jobSummaries: JobSummary[];
}

export type ListTimesheetsParams = {
  /**
   * @nullable
   */
  workerId?: number | null;
  /**
   * @nullable
   */
  worksiteId?: number | null;
  /**
   * ISO date string for start of week filter
   * @nullable
   */
  weekStart?: string | null;
  /**
   * ISO date string for end of week filter
   * @nullable
   */
  weekEnd?: string | null;
};

export type ListMessagesParams = {
  /**
   * @nullable
   */
  toWorkerId?: number | null;
  /**
   * @nullable
   */
  fromWorkerId?: number | null;
};