# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShakaSite Construction Timesheet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job with milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/milestones:
    get:
      operationId: listJobMilestones
      tags: [jobs]
      summary: List milestones for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createJobMilestone
      tags: [jobs]
      summary: Create a milestone for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Milestone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheet entries
      parameters:
        - name: workerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: worksiteId
          in: query
          schema:
            type: ["integer", "null"]
        - name: weekStart
          in: query
          schema:
            type: ["string", "null"]
            description: ISO date string for start of week filter
        - name: weekEnd
          in: query
          schema:
            type: ["string", "null"]
            description: ISO date string for end of week filter
      responses:
        "200":
          description: List of timesheet entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimesheetEntry"
    post:
      operationId: createTimesheetEntry
      tags: [timesheets]
      summary: Create a timesheet entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimesheetEntryBody"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEntry"
  /timesheets/{id}:
    get:
      operationId: getTimesheetEntry
      tags: [timesheets]
      summary: Get a timesheet entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTimesheetEntry
      tags: [timesheets]
      summary: Update a timesheet entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimesheetEntryBody"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEntry"
    delete:
      operationId: deleteTimesheetEntry
      tags: [timesheets]
      summary: Delete a timesheet entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Entry deleted
  /timesheets/{id}/signoff:
    post:
      operationId: signOffTimesheetEntry
      tags: [timesheets]
      summary: Sign off a timesheet entry (manager action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignOffBody"
      responses:
        "200":
          description: Entry signed off
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEntry"
  /timesheets/week/signoff:
    post:
      operationId: signOffWeek
      tags: [timesheets]
      summary: Sign off all entries for a worker's week
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignOffWeekBody"
      responses:
        "200":
          description: Week signed off
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimesheetEntry"
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages
      parameters:
        - name: toWorkerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: fromWorkerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Job:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        worksiteId:
          type: integer
        budgetedHours:
          type: ["number", "null"]
        hoursUsed:
          type: number
        deadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          description: "active, completed, on_hold"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - worksiteId
        - hoursUsed
        - status
        - createdAt
    JobDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        worksiteId:
          type: integer
        budgetedHours:
          type: ["number", "null"]
        hoursUsed:
          type: number
        deadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - worksiteId
        - hoursUsed
        - status
        - milestones
        - createdAt
    CreateJobBody:
      type: object
      properties:
        name:
          type: string
        worksiteId:
          type: integer
        budgetedHours:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
      required:
        - name
        - worksiteId
    UpdateJobBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        budgetedHours:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: ["string", "null"]
    Milestone:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        hoursAtMilestone:
          type: ["number", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        sortOrder:
          type: integer
      required:
        - id
        - jobId
        - name
        - sortOrder
    CreateMilestoneBody:
      type: object
      properties:
        name:
          type: string
        hoursAtMilestone:
          type: ["number", "null"]
        sortOrder:
          type: integer
      required:
        - name
        - sortOrder
    TimesheetEntry:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        worksiteId:
          type: integer
        jobId:
          type: ["integer", "null"]
        date:
          type: string
          description: ISO date YYYY-MM-DD
        startTime:
          type: ["string", "null"]
          description: "HH:MM format"
        finishTime:
          type: ["string", "null"]
          description: "HH:MM format"
        lunchMinutes:
          type: ["integer", "null"]
        totalHours:
          type: ["number", "null"]
        taskBreakdown:
          type: ["string", "null"]
          description: JSON string of task type hours
        notes:
          type: ["string", "null"]
        isSignedOff:
          type: boolean
        signedOffBy:
          type: ["integer", "null"]
        signedOffAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workerId
        - worksiteId
        - date
        - isSignedOff
        - createdAt
        - updatedAt
    CreateTimesheetEntryBody:
      type: object
      properties:
        workerId:
          type: integer
        worksiteId:
          type: integer
        jobId:
          type: ["integer", "null"]
        date:
          type: string
        startTime:
          type: ["string", "null"]
        finishTime:
          type: ["string", "null"]
        lunchMinutes:
          type: ["integer", "null"]
        totalHours:
          type: ["number", "null"]
        taskBreakdown:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - workerId
        - worksiteId
        - date
    UpdateTimesheetEntryBody:
      type: object
      properties:
        startTime:
          type: ["string", "null"]
        finishTime:
          type: ["string", "null"]
        lunchMinutes:
          type: ["integer", "null"]
        totalHours:
          type: ["number", "null"]
        taskBreakdown:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
    SignOffBody:
      type: object
      properties:
        signedOffBy:
          type: integer
      required:
        - signedOffBy
    SignOffWeekBody:
      type: object
      properties:
        workerId:
          type: integer
        weekStart:
          type: string
        signedOffBy:
          type: integer
      required:
        - workerId
        - weekStart
        - signedOffBy
    Message:
      type: object
      properties:
        id:
          type: integer
        fromWorkerId:
          type: integer
        toWorkerId:
          type: integer
        body:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fromWorkerId
        - toWorkerId
        - body
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        fromWorkerId:
          type: integer
        toWorkerId:
          type: integer
        body:
          type: string
      required:
        - fromWorkerId
        - toWorkerId
        - body
...