# lib/api-zod/src/generated/types/createTimesheetEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShakaSite Construction Timesheet API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTimesheetEntryBody {
  workerId: number;
  worksiteId: number;
  /** @nullable */
  jobId?: number | null;
  date: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  finishTime?: string | null;
  /** @nullable */
  lunchMinutes?: number | null;
  /** @nullable */
  totalHours?: number | null;
  /** @nullable */
  taskBreakdown?: string | null;
  /** @nullable */
  notes?: string | null;
}