# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShakaSite Construction Timesheet API
 * OpenAPI spec version: 0.1.0
 */
import type { Milestone } from "./milestone";

export interface JobDetail {
  id: number;
  name: string;
  worksiteId: number;
  /** @nullable */
  budgetedHours?: number | null;
  hoursUsed: number;
  /** @nullable */
  deadline?: Date | null;
  status: string;
  milestones: Milestone[];
  createdAt: Date;
}